requirements.txt
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
argcomplete==3.6.1
backoff==2.2.1
boto3==1.37.23
botocore==1.37.23
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cohere==5.14.0
colorama==0.4.6
Deprecated==1.2.18
distro==1.9.0
eval_type_backport==0.2.2
executing==2.2.0
fastavro==1.10.0
filelock==3.18.0
fsspec==2025.3.1
google-auth==2.38.0
googleapis-common-protos==1.70.0
griffe==1.7.1
groq==0.20.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.0
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.1.0
jiter==0.9.0
jmespath==1.0.1
langfuse==2.60.3
logfire==3.14.0
logfire-api==3.12.0
markdown-it-py==3.0.0
mcp==1.6.0
mdurl==0.1.2
mistralai==1.6.0
nest-asyncio==1.6.0
openai==1.69.0
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-http==1.32.1
opentelemetry-instrumentation==0.53b1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
packaging==24.2
pluggy==1.5.0
prompt_toolkit==3.0.50
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.1
pydantic-ai==0.0.46
pydantic-ai-slim==0.0.46
pydantic-graph==0.0.46
pydantic-settings==2.8.1
pydantic_core==2.33.0
Pygments==2.19.1
pytest==8.3.5
pytest-json-report==1.5.0
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
rsa==4.9
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.46.1
tokenizers==0.21.1
tqdm==4.67.1
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
fastapi==0.110.0
wcwidth==0.2.13
wrapt==1.17.2
zipp==3.21.0
# For crawlers
crawl4ai
openai
# supabase
supabase>=2.4.0

# For image processing
Pillow>=10.0.0  # PIL for image handling and validation
chromadb>=0.4.0  # Vector database for storing embeddings

# For knowledge graph
neo4j>=5.0.0  # Graph database driver

# For calendar parsing
pytz>=2023.3  # Timezone handling for calendar events
